# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MDR App",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings/site-cache/profiles": {
      "get": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "List Profiles",
        "operationId": "list_profiles_api_v1_settings_site_cache_profiles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/site-cache/profiles/upsert": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Upsert Profile",
        "operationId": "upsert_profile_api_v1_settings_site_cache_profiles_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/profiles/delete": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Delete Profile",
        "operationId": "delete_profile_api_v1_settings_site_cache_profiles_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheProfileDeleteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/cidrs": {
      "get": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "List Profile Cidrs",
        "operationId": "list_profile_cidrs_api_v1_settings_site_cache_cidrs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/site-cache/cidrs/upsert": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Upsert Profile Cidr",
        "operationId": "upsert_profile_cidr_api_v1_settings_site_cache_cidrs_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheCIDRIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/cidrs/delete": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Delete Profile Cidr",
        "operationId": "delete_profile_cidr_api_v1_settings_site_cache_cidrs_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheCIDRDeleteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/rules": {
      "get": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "List Pin Rules",
        "operationId": "list_pin_rules_api_v1_settings_site_cache_rules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/site-cache/rules/upsert": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Upsert Pin Rule",
        "operationId": "upsert_pin_rule_api_v1_settings_site_cache_rules_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheRuleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/rules/delete": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Delete Pin Rule",
        "operationId": "delete_pin_rule_api_v1_settings_site_cache_rules_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheRuleDeleteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/tokens": {
      "get": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "List Agent Tokens",
        "operationId": "list_agent_tokens_api_v1_settings_site_cache_tokens_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Profile Id"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/site-cache/tokens/mint": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Mint Agent Token",
        "operationId": "mint_agent_token_api_v1_settings_site_cache_tokens_mint_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheTokenMintIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/tokens/revoke": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Revoke Agent Token",
        "operationId": "revoke_agent_token_api_v1_settings_site_cache_tokens_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheTokenRevokeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/site-cache/rebuild-pins": {
      "post": {
        "tags": [
          "settings-site-cache"
        ],
        "summary": "Rebuild Site Pins",
        "operationId": "rebuild_site_pins_api_v1_settings_site_cache_rebuild_pins_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCacheRebuildPinsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SiteCacheCIDRDeleteIn": {
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SiteCacheCIDRDeleteIn"
      },
      "SiteCacheCIDRIn": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "profile_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Profile Id"
          },
          "cidr": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Cidr"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "cidr"
        ],
        "title": "SiteCacheCIDRIn"
      },
      "SiteCacheProfileDeleteIn": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "hard_delete": {
            "type": "boolean",
            "title": "Hard Delete",
            "default": false
          }
        },
        "type": "object",
        "title": "SiteCacheProfileDeleteIn"
      },
      "SiteCacheProfileIn": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "project_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Code"
          },
          "local_root_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Root Path"
          },
          "fallback_mode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Mode",
            "default": "local_first"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "SiteCacheProfileIn"
      },
...
      "SiteCacheRuleDeleteIn": {
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SiteCacheRuleDeleteIn"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}